Removed the builders, it was pure madness.
We'll use a Serive Locator, which is static (using lazy_static macro).
This service locator will take ownership of our systems... bu they need to
be Sync. The goal was to use concurrency to deal with our various systems,
so we may need to create Arc<Mutex<OurSystem>> in order to use them

Started to add diagrams. Work in progress.